Create dynamic menu for NIC config

+++ frames/InitFrame.py
@@ -4,11 +4,11 @@ from tkinter import messagebox
 
 import gui
 from frames import MainFrame
-from frames import ConfigNCFrame
+from frames import ConfigNICFrame
 
 class InitFrame:
     def __init__(self, master):
-        master.geometry("500x520")
+        master.geometry("500x570")
         frame = tk.Frame(master, bg='white', width=300, height=300)
         frame.place(relx=0.5, rely=0.5, anchor=tk.CENTER)
         
@@ -22,15 +22,15 @@ class InitFrame:
         self.button_border2.grid(column=0, row=1,pady=20)
         self.button2.grid(column=0, row=1)
 
-        self.button_border3, self.button3 = gui.create_button(frame, "Configure network cards")
+        self.button_border3, self.button3 = gui.create_button(frame, "Configure NICs")
         self.button3.configure(command = lambda: self.configure_nc_cards(master))
         self.button_border3.grid(column=0, row=2,pady=20)
-        self.button3.grid(column=0, row=1)
+        self.button3.grid(column=0, row=2)
 
         self.button_border4, self.button4 = gui.create_button(frame, "Exit")
         self.button4.configure(command = lambda: self.exit_program(master))
         self.button_border4.grid(column=0, row=3,pady=20)
-        self.button4.grid(column=0, row=2)
+        self.button4.grid(column=0, row=3)
 
         self.buttons_list = [self.button1,
                             self.button2,
@@ -44,15 +44,15 @@ class InitFrame:
 
     # Go to main configuration frame   
     def main_menu(self, master):
-        gui.check_system(master)
-        gui.destroy_items(self.buttons_list, self.buttons_border_list)
-        e = MainFrame.MainFrame(master)
+        if(gui.check_system(master)):
+            gui.destroy_items(self.buttons_list, self.buttons_border_list)
+            e = MainFrame.MainFrame(master)
     
     # Go to configuration NC cards frame
     def configure_nc_cards(self, master):
-        gui.check_system(master)
-        gui.destroy_items(self.buttons_list, self.buttons_border_list)
-        e = ConfigNCFrame.ConfigNCFrame(master)
+        if (gui.check_system(master)):
+            gui.destroy_items(self.buttons_list, self.buttons_border_list)
+            e = ConfigNICFrame.ConfigNICFrame(master)
     
     # Exit program function
     def exit_program(self, master):
